Patch 3 gamecontrol synchronized

+++ engine/src/main/kotlin/games/figureit/engine/first/MapControl.kt
@@ -1,10 +1,9 @@
 package games.figureit.engine.first
 
 import games.figureit.engine.model.Move
-import games.figureit.engine.model.Player
 import games.figureit.engine.model.Size
 
 interface MapControl {
-    fun move(playerId: Int, move: Move)
+    fun move(playerId: Long, move: Move)
     fun getMapSize(): Size
 }
